requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-environ==0.4.5
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.6
gunicorn==20.0.4
